refactor(transforms): reorganize directory structure for better maintainability
- Move core/transform-utils.ts → utils.ts (remove unnecessary nesting)
- Move pipelines/main.ts → index.ts (single entry point)
- Update all import paths throughout the codebase
- Remove empty core/ and pipelines/ directories
- Update README.md to reflect new structure
- Fix TypeScript import path in transform-core.ts
- All tests passing after reorganization

New structure:
src/transforms/
├── index.ts                    # Main pipeline and exports
├── utils.ts                    # Shared utilities
├── transforms/                 # Individual implementations
└── __tests__/                  # Test files

Benefits:
- Simpler navigation with less nesting
- Clear single entry point (index.ts)
- Better import paths
- More maintainable structure
