+++ packages/web-ui/src/cli/core/shared/transform-core.ts
@@ -67,7 +67,7 @@ export const executeTransforms = async (
 const executeTransformPipeline = async (config: TransformConfig): Promise<TransformResult> => {
   try {
     // Import the main pipeline (dependency injection pattern)
-    const { runMainPipeline } = await import('../../../transforms/pipelines/main.js');
+    const { runMainPipeline } = await import('../../../transforms/index.js');
 
     // Execute pipeline with configuration
     const result = await runMainPipeline(config.srcDir, {
@@ -80,7 +80,7 @@ const executeTransformPipeline = async (config: TransformConfig): Promise<Transf
       filesProcessed: result.processedFiles,
       filesModified: config.dryRun ? 0 : result.processedFiles,
       conversionsApplied: result.processedFiles * 2, // Estimate: avg 2 conversions per file in main pipeline
-      errors: result.errors.map(e => e.error),
+      errors: result.errors.map((e: any) => e.error),
       warnings: [],
     };
   } catch (error) {
